Change Remarks for hw_04
